Collision fixes. Some bugfixes after dt introduction

+++ src/math/Vector2.hpp
@@ -52,12 +52,12 @@ struct Vector2
     Vector2 operator/(const T d)
     {
         return Vector2{x/d, y/d};
-    } 
+    }
 
 };
 
 using Vector2i = Vector2<int32_t>;
-using Vector2f = Vector2<float>;
+using Vector2f = Vector2<double>;
 
 }  // namespace math
